+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Fixed
 - MerchantCategoryNumber type in IssuingPurchase resource, the API returns a string
 - BusinessAccountRequest resource
+- IndividualAccountRequest resource
+- IndividualAccountRequestLog resource
+- IndividualAccountAttachment resource
+- IndividualAccountAttachmentLog resource
 
 ## [1.2.0] - 2026-07-03
 ### Fixed